fix(pr-review): resolve compile errors and add sqlx cache
Auto-Fix Changes (iteration #3):
- Fix migration FK reference: tenants(id) -> tenants(tenant_id)
- Fix SUM type mismatch: cast to BIGINT for i64 return type
- Regenerate sqlx offline cache with new landed costs queries

Quality Gates Passed:
- cargo fmt --check: OK
- cargo clippy (inventory_service): OK
- cargo test --lib: 153 passed

+++ services/inventory_service/infra/src/repositories/landed_cost.rs
@@ -676,7 +676,7 @@ impl LandedCostLineRepository for LandedCostLineRepositoryImpl {
     async fn sum_lines_amount(&self, tenant_id: Uuid, document_id: Uuid) -> Result<i64, AppError> {
         let result = sqlx::query!(
             r#"
-            SELECT COALESCE(SUM(amount), 0) as "total!"
+            SELECT COALESCE(SUM(amount)::BIGINT, 0) as "total!"
             FROM landed_cost_lines
             WHERE tenant_id = $1 AND document_id = $2
             "#,
